fixed landmarks position calculation for fiber length input by user

+++ Shared/Utils35/SorData/SorData.cs
@@ -101,12 +101,12 @@ namespace Iit.Fibertest.UtilsLib
         }
 
 
-        public static int GetDistanceBetweenLandmarksInMm(
+        public static double GetDistanceBetweenLandmarksInMm(
             this OtdrDataKnownBlocks sorData, int leftIndex, int rightIndex)
         {
             var owt1 = sorData.LinkParameters.LandmarkBlocks[leftIndex].Location;
             var owt2 = sorData.LinkParameters.LandmarkBlocks[rightIndex].Location;
-            return (int)((owt2 - owt1) * GetOwtToMmCoeff(sorData));
+            return (owt2 - owt1) * GetOwtToMmCoeff(sorData);
         }
 
         public static int GetOwtFromMm(this OtdrDataKnownBlocks sorData, int distance)
